# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API Antigüedades Sthandier",
    "version": "1.0.0",
    "description": "Documentación para el CRUD de productos, reservas y usuarios"
  },
  "servers": [
    {
      "url": "http://localhost:3000/api"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "Usuario": {
        "type": "object",
        "required": [
          "nombre",
          "apellido",
          "email",
          "contraseña",
          "rol"
        ],
        "properties": {
          "id": {
            "type": "integer",
            "example": 1
          },
          "nombre": {
            "type": "string",
            "example": "Juan"
          },
          "apellido": {
            "type": "string",
            "example": "Pérez"
          },
          "email": {
            "type": "string",
            "format": "email",
            "example": "juan@example.com"
          },
          "contraseña": {
            "type": "string",
            "format": "password",
            "example": "password123"
          },
          "rol": {
            "type": "string",
            "enum": [
              "admin",
              "cliente",
              "empleado"
            ],
            "example": "cliente"
          },
          "telefono": {
            "type": "string",
            "example": "+56912345678"
          },
          "direccion": {
            "type": "string",
            "example": "Av. Principal 123"
          },
          "estado": {
            "type": "string",
            "enum": [
              "Activo",
              "Inactivo"
            ],
            "example": "Activo"
          }
        }
      }
    }
  },
  "paths": {
...
    "/api/usuarios": {
      "get": {
        "summary": "Obtener todos los usuarios",
        "tags": [
          "Usuarios"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de usuarios",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Usuario"
                  }
                }
              }
            }
          },
          "401": {
            "description": "No autorizado"
          },
          "404": {
            "description": "No se encontraron usuarios"
          }
        }
      },
      "post": {
        "summary": "Crear un nuevo usuario",
        "tags": [
          "Usuarios"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Usuario"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Usuario creado exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Usuario creado exitosamente."
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Datos inválidos"
          },
          "409": {
            "description": "El correo electrónico ya está registrado"
          }
        }
      }
    },
    "/api/usuarios/{id}": {
      "get": {
        "summary": "Obtener un usuario por ID",
        "tags": [
          "Usuarios"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "integer"
            },
            "required": true,
            "description": "ID del usuario"
          }
        ],
        "responses": {
          "200": {
            "description": "Usuario encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Usuario"
                }
              }
            }
          },
          "401": {
            "description": "No autorizado"
          },
          "404": {
            "description": "Usuario no encontrado"
          }
        }
      }
    },
    "/api/usuarios/{email}": {
      "put": {
        "summary": "Actualizar un usuario",
        "tags": [
          "Usuarios"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "email",
            "schema": {
              "type": "string",
              "format": "email"
            },
            "required": true,
            "description": "Email del usuario a actualizar"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Usuario"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Usuario actualizado",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Usuario actualizado exitosamente."
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "No autorizado"
          },
          "404": {
            "description": "Usuario no encontrado"
          }
        }
      },
      "delete": {
        "summary": "Eliminar un usuario",
        "tags": [
          "Usuarios"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "email",
            "schema": {
              "type": "string",
              "format": "email"
            },
            "required": true,
            "description": "Email del usuario a eliminar"
          }
        ],
        "responses": {
          "200": {
            "description": "Usuario eliminado",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Usuario eliminado exitosamente."
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "No autorizado"
          },
          "404": {
            "description": "Usuario no encontrado"
          }
        }
      }
    },
    "/api/usuarios/desactivar": {
      "post": {
        "summary": "Desactivar un usuario (solo admin)",
        "tags": [
          "Usuarios"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email",
                    "example": "admin@example.com"
                  },
                  "contraseña": {
                    "type": "string",
                    "format": "password",
                    "example": "admin123"
                  },
                  "email_modificar": {
                    "type": "string",
                    "format": "email",
                    "example": "usuario@example.com"
                  },
                  "estado": {
                    "type": "string",
                    "enum": [
                      "Activo",
                      "Inactivo"
                    ],
                    "example": "Inactivo"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Estado del usuario actualizado",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Usuario usuario@example.com actualizado exitosamente."
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "No autorizado o credenciales incorrectas"
          },
          "403": {
            "description": "Usuario no tiene permisos de administrador"
          },
          "404": {
            "description": "Usuario no encontrado"
          }
        }
      }
    }
  },
...
}